added account page with basic structure
