+++ src/app/(pages)/channel/[...id]/components/postInput.jsx
