+++ src/app/(pages)/channel/[...id]/components/posts.jsx
