+++ src/app/components/loading/loadingComponent.jsx
@@ -1,0 +1,6 @@
+const LoadingComponent = () => {
+  return (
+    <p>Loading...</p>
+  );
+}
+export default LoadingComponent;
