+++ src/app/page.js
@@ -2,8 +2,6 @@
 
 import Landing from "./components/pages/landing/page"
 
-import { redirect } from 'next/navigation'
-
 export default async function Home() {
 
   return (
